requirements.txt
blinker==1.4
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
future==0.18.2
idna==2.10
oauthlib==3.0.1
pip==20.2.3
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dotenv==0.14.0
python-twitter==3.5
requests==2.24.0
requests-oauthlib==1.3.0
setuptools==49.6.0.post20200925
six==1.15.0
urllib3==1.25.10
wheel==0.35.1
